+++ lib/pymedphys/_experimental/pseudonymisation/api.py
@@ -1,7 +1,7 @@
 # pylint: disable = unused-import
 
 
-from pymedphys._experimental.pseudonymisation import (
+from pymedphys._experimental.pseudonymisation.main import (
     get_default_pseudonymisation_keywords,
     is_valid_strategy_for_keywords,
     pseudonymise,
